+++ ExilenceNextApp/src/store/domains/profile.ts
@@ -252,11 +252,6 @@ export class Profile {
   }
 
   @action
-  setActiveStashTabs(stashTabIds: string[]) {
-    this.activeStashTabIds = stashTabIds;
-  }
-
-  @action
   updateFromApiProfile(apiProfile: IApiProfile) {
     this.activeLeagueId = apiProfile.activeLeagueId;
     this.activePriceLeagueId = apiProfile.activePriceLeagueId;
@@ -299,7 +294,7 @@ export class Profile {
     visitor!.event('Profile', 'Triggered snapshot').send();
 
     rootStore.uiStateStore!.setIsSnapshotting(true);
-    this.getItems();
+    this.refreshStashTabs();
   }
 
   @action clearIncome() {
@@ -352,6 +347,49 @@ export class Profile {
       rootStore.accountStore.getSelectedAccount.queueSnapshot();
     }
     rootStore.uiStateStore!.setIsSnapshotting(false);
+  }
+
+  @action refreshStashTabs() {
+    const accountLeague = rootStore.accountStore.getSelectedAccount.accountLeagues.find(
+      (al) => al.leagueId === this.activeLeagueId
+    );
+
+    const league = rootStore.leagueStore.leagues.find((l) => l.id === this.activeLeagueId);
+
+    if (!accountLeague || !league) {
+      return this.getItemsFail(new Error('no_matching_league'), this.activeLeagueId);
+    }
+
+    rootStore.uiStateStore.setStatusMessage('refreshing_stash_tabs');
+
+    fromStream(
+      accountLeague.getStashTabs().pipe(
+        mergeMap(() => of(this.refreshStashTabsSuccess(league.id))),
+        catchError((e: AxiosError) => of(this.refreshStashTabsFail(e, league.id)))
+      )
+    );
+  }
+
+  @action refreshStashTabsSuccess(leagueId: string) {
+    rootStore.notificationStore.createNotification(
+      'refreshing_stash_tabs',
+      'success',
+      undefined,
+      undefined,
+      leagueId
+    );
+    this.getItems();
+  }
+
+  @action refreshStashTabsFail(e: AxiosError | Error, leagueId: string) {
+    rootStore.notificationStore.createNotification(
+      'refreshing_stash_tabs',
+      'error',
+      true,
+      e,
+      leagueId
+    );
+    this.snapshotFail();
   }
 
   @action getItems() {