+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ##### Added
 - Added style tweaks to the item table
 - Added persistence of position/size for the overlay
+- Added reindexing of stash tabs before every snapshot, to support moving tabs around
 ##### Fixed
 - Fixed a bug where auto snapshotting would sometimes stop working
 - Fixed a bug where the net worth overlay would not update when income was reset